Fix package name based on local sdist archive
When using a local sdist as source, the package name of the conda recipe
was based on the sdist filename.
This used to be the same but since setuptools 69.3.0, the sdist filename is now normalized.
For example, if a package name contains a dash, it's converted to underscore.

The package name should be extracted from the metadata and not the sdist filename.