Retrieve packages from top_level.txt when possible
The "packages" key that contains the list of modules part of the distribution
is used in the test imports section of the conda recipe.

It can be retrieved from setup.py.

When only pyproject.toml is present, this is usually not defined.
It can be added but depends on the build backend.

For project using setuptools, this information can be retrieved from the top_level.txt
file created in the .egg.info directory.

When "packages" isn't present, grayskull will use the name of the distribution,
but that might differ.