Remove support for python 3.8 (#579)
* Remove support for python 3.8

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* Update type annotations to Python 3.9+ syntax

Replace legacy type annotations with Python 3.9+ syntax for improved readability and conciseness. Additionally, update the pre-commit hook version for ruff to v0.7.3 for linting purposes.

Signed-off-by: Marcelo Trevisani <marceloduartetrevisani@gmail.com>

* Update Python version in GHA workflows

Set Sphinx workflow to use Python 3.12 explicitly and removed testing for Python 3.9 in the tests workflow.

Signed-off-by: Marcelo Trevisani <marceloduartetrevisani@gmail.com>

* Update CLIConfig type hint for singleton instance

Replaced the type hint for the `__instance` attribute in `CLIConfig` from `"CLIConfig"` to `Self` to improve type checking and readability.

Signed-off-by: Marcelo Trevisani <marceloduartetrevisani@gmail.com>

* Handle missing 'Self' in typing import

Added a fallback mechanism to import 'TypeVar' and define 'Self' if 'Self' is not available in the typing module. This ensures compatibility with older Python versions that do not support 'Self'.

Signed-off-by: Marcelo Trevisani <marceloduartetrevisani@gmail.com>

---------

Signed-off-by: Marcelo Trevisani <marceloduartetrevisani@gmail.com>
Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>