Add several pypi tests
These tests are designed to make sure multiple selectors for the
same package are preserved correctly at every stage of the process
of merging dependencies from different sources.

When parsing poetry dependencies, we introduce selectors early
in the process, whereas for other approaches, we retain "extras"
(separated by semicolons) for much of the merging process. Packages
employing each strategy are used in the added tests.